+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 4.8.0
+## 1.2.1
+
+- Added new methods sendInAppImpression and sendInAppClick to send impression or/and click for nativeAd
+- Added new method openNativeAd to execute CTA action (open webview on browser or inapp)
+
+## 1.2.0
 
 - Android X migration
 - iOS 14 optimized